Recast config and challenge result as issuer
Moves config and challenge result into legacy issuer package. API
servers now use issuer pool containing the legacy issuer.
